chore: add tree view for goals list
based on: https://iamkate.com/code/tree-views/
