+++ app/[id]/components/goal-list/goal-list.module.css
@@ -1,0 +1,76 @@
+.GoalTree {
+  --spacing: 1.5rem;
+  --radius: 10px;
+}
+
+.GoalTree li {
+  display: block;
+  position: relative;
+  padding-left: calc(2 * var(--spacing) - var(--radius) - 2px);
+  padding-top: var(--radius);
+}
+
+.GoalTree ul {
+  margin-left: calc(var(--radius) - var(--spacing));
+  padding-left: 0;
+}
+
+.GoalTree ul li {
+  border-left: 2px solid #ddd;
+}
+
+.GoalTree ul li:last-child {
+  border-color: transparent;
+}
+
+.GoalTree ul li::before {
+  content: "";
+  display: block;
+  position: absolute;
+  top: calc(var(--spacing) / -2);
+  left: -2px;
+  width: calc(var(--spacing) + 2px);
+  height: calc(var(--spacing) + 1px);
+  border: solid #ddd;
+  border-width: 0 0 2px 2px;
+}
+
+.GoalTree summary {
+  display: block;
+  cursor: pointer;
+}
+
+.GoalTree summary::marker,
+.GoalTree summary::-webkit-details-marker {
+  display: none;
+}
+
+.GoalTree summary:focus {
+  outline: none;
+}
+
+.GoalTree summary:focus-visible {
+  outline: 1px dotted #000;
+}
+
+.GoalTree li::after,
+.GoalTree summary::before {
+  content: "";
+  display: block;
+  position: absolute;
+  top: calc(var(--spacing) / 2 - var(--radius));
+  left: calc(var(--spacing) - var(--radius) - 1px);
+  width: calc(2 * var(--radius));
+  height: calc(2 * var(--radius));
+  border-radius: 50%;
+  background: #ddd;
+}
+
+.GoalTree summary::before {
+  z-index: 1;
+  background: #696 url("/expand-collapse.svg") 0 0;
+}
+
+.GoalTree details[open] > summary::before {
+  background-position: calc(-2 * var(--radius)) 0;
+}
